Fix Makefile dependencies for js/decoder.js
js/decoder-post.js was an order-only dependency; this doesn't make much
sense for files that aren't generated by other rules, so changing that
file didn't force js/decoder.js to rebuild.

What we need to do instead is to specify this file in the list of all
prerequisites, and only feed .cpp files from that list to emcc for
compilation.