indexcodec: Use C-style casts for integer casts
The convention established in this code is to use C++ casts for pointers
and functional casts or C casts if the type has a space for integers. We
were mostly consistent wrt that with a few minor exceptions in
indexcodec.

Once we switch to stdint.h, we'll be able to replace all unsigned chars
with uint8_t and functional casts...