simplify: Implement binary search for the correct cell size
We now compute a target cell count based on the approximate 2x ratio
between triangles and vertices and then binary search through the grid
cell sizes to find the right one.

This substantially reduces simplification performance, however we should
be able to try to guess the optimal cell size faster than 10 steps. An
open question is whether we need to do precise counting of triangles
along the way - right now we don't do that which means we don't actually
guarantee that the output is going to be under target triangle count...