simplify: Add a binary-only cell construction mode
This mode is notably faster to find an approximation for cell count, but
it's much less precise - it only can step the grid size 2x.

However, this may work nicely with an adaptive cell merging method. When
the first pass generates a somewhat overly fine grid, the second pass
can attempt to clean up the results by adaptively merging the cells.