simplify: Explore a simpler version of a grid size search
Instead of doing custom binary passes + reduce passes + upres passes,
this change uses a binary search, augmented with initial guess and
interpolations search to accelerate it.

We may need to tweak interpolation search to make sure that the process
converges in 10 steps, but in general this seems to work reasonably well
- on par with the custom passes but with less code. Still not 100% sure
what we need to settle on.