simplify: Use combined interpolation+binary search
We now use a few interpolation passes and then fall back to binary
search. We don't limit the number of passes as a result - we're
guaranteed to converge because binary search computes a bit for grid
size per pass so we're going to converge in 10 binary passes.

For now settle on interpolation search without a guard criteria and 5
interpolation passes. We can tweak this in the future.

This seems to be on par or better performance-wise compared to our old
custom method, and requires significantly less code, so let's just do
this!