js: Use sbrk instead of malloc/free
We don't use malloc/free in the vertex/index decoders, we only need them
to allocate memory for source/target buffers from JS. This allocation is
stack-like so it's sufficient to use sbrk() for that.

As a result, we don't need emmalloc (which is ~2KB wasm which translates
to ~2.7KB after Base64 encoding).

This change reduces js/decoder.js from 19.7 KB to 15.4 KB before gzip,
and from 8.4 KB to 6.7 KB after gzip. Some of the savings come from
upgrading to newer emscripten version, but most of this is from removing
emmalloc.