simplify: Improve simplifySloppy performance for large reductions
We were previously using generic fillFaceQuadrics that assumed that a
triangle contributes to 3 quadrics; this is true for default simplifier,
and true for sloppy simplifier when reduction is not very significant,
but for large simplifications most triangles contribute to just one
cell.

We can leverage this by pre-scaling resulting quadric by 3 and only
adding it once, which saves 20 FLOPs for most quadrics at the cost of a
branch.

As a result, large (10x-100x) reductions are now between 10% and 20%
faster.