simplify: Add multiple different modes of operation
We now have an easy way to experiment with different quadric weighting
options, quadric ranking/limiting modes (although the limit is inactive
in lodviewer atm) and memoryless vs memorized quadrics.

This should make it easier to tune the error metric on real meshes.

Note that A in lodviewer only switches between unnormalized error and
normalized error for rank & limit since we don't use limit in lodviewer
with this change.