demo: Implement summary analysis and distance error metric
We now average errors across all tested meshes and print the errors at
the end; this allows us to more easily compare the results in bulk.

Also, in addition to just counting number of mismatching pixels, we
compute the distance metric - Manhattan distance to the closest pixel on
the other image. This results in more robust error measurement.
(obviously Euclidean distance would be more correct but Manhattan is
easier to compute).