simplify: Rework error metric to use normalized distance
Several experiments were performed in order to establish a good metric:

- How do we weight quadrics? weight = 1, sqrt(area), area, area^2 were
evaluated. area^2 is commonly used in literature since it results in
area^2*distance^2 = volume^2 resulting error from the quadric, and
minimizing that minimizes volume.

- How do we rank collapses? Before this work, we used unnormalized error
- this is what all the literature uses but it results in errors that are
incredibly hard to reason about since they incorporate scaling factors
that depend on local geometry. Possibilities of using unnormalized error
for ranking & limiting, using unnormalized error for ranking but
limiting collapses based on normalized error, or using normalizer error
for both were evaluted.

- Do we store quadrics between passes? "Memoryless" simplification can
be emulated by recomputing quadrics for all faces/edges for each pass.

As a result, a few things became obvious.

It's absolutely critical to limit collapses based on normalized error.
This is the only way to be able to give consistent results independent
of model complexity. With this, it becomes possible to specify the input
error, and simplify to 0 triangles - and many different meshes result in
roughly equal complexity irrespective of the original triangle count.

Because of this, we now keep accumulated weight in each quadric.

Volume preservation is not as important as silhouette preservation.
area^2 weight performs relatively poorly because of this. A corner case
is a tetrahedron with a long but very thin base - simplifying that is
good from perspective of any weighting scheme that uses area, but it
produces objectionable artifacts on silhouettes.

Because of this, we now use linear weighting for all metrics. Area
weighting performs reasonably well but not as good comparatively.

Memoryless simplification, contrary to the literature, was inferior to
simplification schemes that preserved quadrics *if* normalized errors
were used. I'm attributing this to the fact that in literature, quadric
weights aren't preserved throughout simplification, which distorts the
meaning of quadrics (since weight keeps growing) in areas that were heavily
simplified.

Because of this, we don't consider memoryless simplification; in
addition to having inferior quality, it's slower - due to the structure of
our algorithm, reusing errors is beneficial and has no cost.

Before this decision was made, a rasterization-based automatic error
measure was implemented; using orthographic renders of a model, the
error was computed as a sum of distances between a solid pixel in
original image and closest pixel in simplified image and vice versa.
This provides a reasonable approximation of image-based error.

The results for each of the 4 weighting schemes were as follows, where
error rate is the percentage of different pixels and error is the
aforementioned sum of distances.

Memoryless:
1:          error rate 8.89%, error 6.831430e-01
sqrt(area): error rate 13.49%, error 9.407884e-01
area:       error rate 17.96%, error 1.230729e+00
area^2:     error rate 23.66%, error 1.684588e+00

Baseline, unnormalized error:
1:          error rate 8.75%, error 8.617142e-01
sqrt(area): error rate 13.19%, error 6.436804e-01
area:       error rate 18.08%, error 9.941355e-01
area^2:     error rate 23.71%, error 1.698783e+00

Normalized limit & ranking (this change):
1:          error rate 13.63%, error 9.811575e-01
sqrt(area): error rate 13.02%, error 6.414265e-01
area:       error rate 12.98%, error 8.759209e-01
area^2:     error rate 14.97%, error 1.355619e+00

Normalized limit, unnormalized ranking:
1:          error rate 14.26%, error 1.039525e+00
sqrt(area): error rate 12.51%, error 6.191866e-01
area:       error rate 12.13%, error 6.591008e-01
area^2:     error rate 12.05%, error 7.602873e-01

It's clear that normalization is necessary to provide a guaranteed
result - note that in both memoryless and memorized quadrics, error rate
varies significantly with the choice of the error metric.

The absolute best option was using sqrt(area) with normalized limit and
unnormalized ranking; however, this option is not as easy to understand
and maintain due to two very different notions of errors. Based on
visual comparison, the difference doesn't seem substantial one way or
the other.