js: Refactor interface to make the library more module-like
Instead of exposing a function that creates a brand new WebAssembly
instance every time - which is prone to abuse - expose a global
MeshoptDecoder object with a .ready Promise and .decode methods.

The object is exposed using the same export tricks Emscripten uses -
hopefully that will make this usable from various JS loaders.