tools/meshencoder: Scale normal so that largest component is +-1
This results in the normal data being more compressible - we save ~1.5
bits per vertex - and being higher precision as well since we
effectively scale each component of the normal up, reducing the
quantization error relative to the original data.

This means that the normals that are fed to the shader aren't normalized
anymore, however this is already the case with ThreeJS that applies
scaling matrix to normal data, and has to renormalize as a result.

The change reduces nyra.optmesh by 3KB (1.5%).