simplify: Reduce the number of passes in some edge cases
We previously would stop the pass when we reached the error goal,
computed as the 1.5 * collapse error #k where k is the number of edges
we need to collapse.

This heuristic was designed to provide a balance between stopping too
early, when we could have collapsed more edges in a single pass, and
collapsing too many edges when the next pass could have edges with
smaller error (since they will become available for collapse due to not
being locked anymore).

On some meshes, this resulted in very large number of passes where each
pass would collapse just a few edges. As an extreme example,
hairball.obj had ? passes.