gltfpack: Cleanly handle multiple texture coordinate sets
We now emit texcoord binding for texture references, and only emit
texture coordinate data that is used by the material.

In some cases some GLTF files have several texture coordinate references
pointing at the same attribute stream; while we could deduplicate this
in the file, with this change it's not very interesting because they
usually wouldn't be used.