stripify: Switch to an explicit restart index
The main motivation behind this is that meshopt_unstripify *needs* an
explicit restart index to be able to work correctly for 16-bit indices.
meshopt_stripify currently relies on 32->16 bit truncation but this
doesn't work in the other direction.

Because of this, it seems cleaner to just settle on passing an explicit
restart index to stripify as well. This also resolves bool/int
complexity in C89 since we don't need a boolean argument anymore :)