gltfpack: Fix single-frame animation truncation
With aggressive resampling, for example 60 -> 15 Hz, we could lose up to
0.1 seconds of animation data previously because of incorrect rounding.

This could result in visible jolts for looping animations.

We now guarantee that we don't lose any data - the last frame will have
the timestamp that either matches the original maxt, or is the earliest
time *after* it. As a result, it's possible that animations will get a
tiny bit of stutter in the end, but that's unavoidable with low
frequency sampling.