gltfpack: Add (disabled) quaternion/keyframe compression
Add code that writes quantized quaternion tracks - this doesn't work in
Three.JS or Babylon.JS but that's due to bugs in these frameworks that
will need to be fixed.

Also add compression using vertex codec for quaternion tracks - it works
really well! - will need to figure out how to cleanly integrate that.