gltfpack: Implement removal of constant animation tracks
We use a heuristic for when constant tracks aren't useful: if *all*
tracks within the file that target the node path have the same value as
the base, this is indicative of the fact that we won't need this data.

This is not definitive - if glTF file *only* stores animation data, you
might want to load this file over some other file with mesh data, and
actually pose the mesh accordingly - but this usecase is unlikely.

This behavior can be disabled using -ac switch which will preserve
constant animation tracks.