gltfpack: Avoid the use of interleaved buffers in three.js
We're currently hitting bugs in three.js associated with their use of
InterleavedBuffer objects and incorrect tracking of array count.

While this isn't a good long-term solution, for now work around this by
using vec4 for positions and normals - this makes sure that byteStride
matches the element size which means that three.js will use
BufferAttribute objects that are supported correctly.