gltfpack: Preserve ancestry of skinned meshes
We're hitting a bug with bounding box generation in three.js where it
uses the world matrix of the node with the mesh to compute the bounds
regardless of whether the mesh is skinned or not.

This can result in substantial differences between the bounding box
before and after gltfpack.

To work around this, we now parent skinned meshes to their original
place in the hierarchy. This requires us to be a bit more generic wrt
the mesh-node mapping handling, which may be helpful in the future.

Contributes to #44 by fixing CesiumMan and BrainStem in three.js.