gltfpack: Fix position min/max in presence of morph targets
When morph targets are used, we extend the quantization range to account
for the possible position change. This is important because three.JS
transforms delta streams into absolute streams, which may overflow the
target range if we aren't careful.

Additionally, this fixes getPositionBounds to compute unsigned vs signed
min/max for base and morph streams respectively, which fixes validation
warnings.