gltfpack: Use uncompressed 8-bit weight animation output
Right now byte stride for compressed attribute streams needs to be
divisible by 4, so weight animation can't be compressed unless it uses
full float32 data.

Additionally, ideally compression would use corellation of weights that
are located in the same positions in the stream, but right now this is
not possible without changes to the attribute decoder.

Weight animation data tends to be minimal so resolve this by not
compressing the data at all, and using 8-bit weights which should really
be enough.