gltfpack: Add compression mode to JSON buffer view declaration
This makes the mode a bit more explicit - we no longer need to rely on
the buffer contents (hypothetically a loader may decide to load buffer
contents in a separate worker without waiting for the first byte to
dispatch the job with the correct type).

This increases the JSON size a bit, but the number of views is generally
low enough to make sure the overhead isn't very significant.