gltfpack: Always preserve node names
The extra complexity of trying to guess whether node names are relevant
doesn't seem justified - in node-heavy scenes typically most nodes are
joints and we keep their names anyway. With this change we preserve all
node names present in the original file - if this ever becomes a size
problem we can add a separate "strip names" flag.

This increases JSON size by ~200-300 bytes for jrat and izzy test models
(out of 32K and 240K respectively). For jrat, the cumulative size increase
taking into account the .glb data is ~0.05% before gzip and ~0.03% after.