gltfpack: Implement material merging and pruning
Often glTF models have many materials that are equivalent. Preserving
these materials wastes space due to extra material specification,
increases the rendering time by increasing the amount of state setup and
also reduces the efficiency of mesh merging.

This change merges materials by retargeting all meshes to use the first
equivalent material in the set, and then removing materials that aren't
referenced anymore from the output.

As a result, in some cases the material count is reduced dramatically;
for example for Buggy.gltf before this change we had 148 output
primitives and 148 materials (and 90 KB of JSON data), and after this
change we get 10 output primitives/materials and 7 KB of JSON data.

Contributes to #48 (indirectly)