gltfpack: Implement support for keeping named nodes
This change implements a flag, -kn, that restricts the mesh merging and
node tree simplification assuming that nodes with a name are important
to preserve and can be manipulated externally.

This amounts to a few things:

- We keep named nodes around even if they are not otherwise referenced.
This can be important for preserving attachment points for characters
for example.

- We keep meshes attached to their original nodes so that manipulating
named nodes transforms the meshes

- We block same-parent mesh merging if one of the nodes is named since
named transforms may change