spatial: Introduce spatialSort and spatialSortTriangles
These functions are designed to sort vertices or triangles by the
spatial proximity. The goal is to improve spatial locality to make
other, subsequently ran algorithms, perform better.

Currently the implementation is using Morton order, which is a
reasonable proxy but in the future we will try to use a kD tree sort.