Use meshopt_Allocator allocators in meshopt_IndexAdapter
Before this change, all allocations inside the library were using the
overrideable allocators, but IndexAdapter was still using global
operator new/delete unconditionally.

To fix this, the functions in the header needed to be reordered; take
this opportunity to more cleanly separate the interface for C++
templated wrappers from implementation.