gltfpack: Implement basic support for simplification
Two extra command-line flags are provided: -si R to simplify the mesh to
a fraction R of its triangle count (e.g. 0.5 to reduce the triangle
count 2x) and -sa to use aggressive simplifier that is guaranteed to
reach the target complexity.

The simplification currently ignores point clouds and needs to become a
bit more adaptive to handle cases when simplifier gets stuck.