js: Update to Emscripten 1.38.47
This change updates the js build to Emscripten 1.38.47. This version
changes the behavior of standalone .wasm modules - instead of importing
a few symbols, including memory, externally, the standard library now
creates and exports the memory object. Because of this the JS embedding
structure needed to change a bit.

Additionally, the ChakraCore workaround is not necessary anymore because
the bug isn't being hit with LLVM upstream codegen.

As a result, the JS is ~2% smaller after these updates.

Note that because of a regression in .47, building this locally requires
patching Emscripten with this change:

https://github.com/emscripten-core/emscripten/pull/9588