gltfpack: Add fallback:true attribute to fallback buffers
@donmccurdy had a great suggestion to mark the dummy buffer in a special
way. This change marks the fallback buffer with fallback:true attribute.

This can be useful for two reasons:

- Fallback buffer without a URI is somewhat odd. The attribute helps
note that this is not a regular buffer and it relates to the compression
extension.

- Fallback attribute is actually helpful for loaders that recognize
MESHOPT_compression - they can skip loading the buffer from the URI.
This is not necessary for Babylon.JS / Three.JS but could help in the
future.

The specification could be something like "if fallback:true is
specified, all buffer views that refer to this buffer must have a
MESHOPT_compression extension override that targets a different buffer".
So specifying this attribute is optional if you want to mix & match
fallback data with actual data in a real buffer.

Then it naturally follows that if the extension support is required,
fallback buffers can be ignored for the purpose of validation etc.