js: Use a feature detector module to pick the wasm binary
We now try to validate a small module that contains bulk memory and SIMD
instructions; if it validates we assume that SIMD is supported and
instantiate it.

Note that we could validate the full wasm_simd module; however, it's
slightly slower (~1ms) to do so, and, more importantly, this can result
in silent issues when browsers have partial implementation of the
instruction set - and I'd rather see the error.