gltfpack: Impement support for Basis texture conversion
When -tb is specified, we run all textures from the input asset through
Basis, and place the results into .basis folder next to the scene asset.

We currently don't distinguish between normal maps and color maps, and
always generate mipmaps (which is a prerequisite for Three.JS Basis
support working). Also there doesn't seem to be a clear path to support
embedding Basis textures atm so we don't support that either.