gltfpack: Implement support for encoding embedded images
If an image is already embedded and we need to encode it with Basis, we
write it to a temporary file (assuming .png for now since basisu doesn't
support jpg...), run Basis file encoder and re-embed the result.

Note that existing viewers don't support image/basis MIME type but
hopefully this can be easily adapted to use KHR_image_ktx2.