gltfpack: Use temporary directory for temporary files
Instead of writing temporary files side by side with the output file, we
write them to the temporary folder instead. This means that if the
process crashes, we don't keep files in the output folder.

Note that the files need a suffix to work around basisu issue where the
output file name has to have a .basis extension. For this to work we
need to use mkstemps on POSIX systems to avoid race conditions.