gltfpack: Fix KTX2 DFD to validate cleanly
Our DFD had incorrect size - it was much larger than it should be. This
change fixes this and also removes all sample information from DFD -
while the spec says it must be included, the validator says it must not
be included and we will trust the validator for now.

In a similar vein, use UNSPECIFIED color model because it validates
cleanly even though the spec disagrees.