vertexcodec: Use a unified cpuid path with caching
cpuid can be moderately expensive - dozens to hundreds of cycles on some
CPUs, although it's likely to only be dozens of cycles for the specific
bits we're interested in.

Additionally, we need to be able to fake the cpuid bits dynamically to
be able to test scalar decoding in unit tests.

To achieve both goals, we now use a unified cpuid-based feature
detection with cpuid cached in meshopt::cpuid global; it's externally
visible to be able to manipulate it in tests.