vcacheoptimizer: Add a separate score table for strip-like sequences
The new score table has been tuned to minimize the encoded index buffer
size. We'll need to re-tune it again after some index codec changes but
it seems to produce longer strips (which are more efficient to compress
post-deflate) so I'm calling this variant "strip".

This also externalizes the table-driven optimizer which will make it
easier to build vcachetuner.